+++ src/Symfony/Component/Validator/Tests/Constraints/UniqueValidatorTest.php
@@ -1,0 +1,89 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Validator\Tests\Constraints;
+
+use Symfony\Component\Validator\Constraints\Unique;
+use Symfony\Component\Validator\Constraints\UniqueValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
+
+class UniqueValidatorTest extends ConstraintValidatorTestCase
+{
+    protected function createValidator()
+    {
+        return new UniqueValidator();
+    }
+
+    /**
+     * @expectedException \Symfony\Component\Validator\Exception\UnexpectedValueException
+     */
+    public function testExpectsUniqueConstraintCompatibleType()
+    {
+        $this->validator->validate('', new Unique());
+    }
+
+    /**
+     * @dataProvider getValidValues
+     */
+    public function testValidValues($value)
+    {
+        $this->validator->validate($value, new Unique());
+
+        $this->assertNoViolation();
+    }
+
+    public function getValidValues()
+    {
+        return [
+            yield 'null' => [[null]],
+            yield 'empty array' => [[]],
+            yield 'single integer' => [[5]],
+            yield 'single string' => [['a']],
+            yield 'single object' => [[new \stdClass()]],
+            yield 'unique booleans' => [[true, false]],
+            yield 'unique integers' => [[1, 2, 3, 4, 5, 6]],
+            yield 'unique floats' => [[0.1, 0.2, 0.3]],
+            yield 'unique strings' => [['a', 'b', 'c']],
+            yield 'unique arrays' => [[[1, 2], [2, 4], [4, 6]]],
+            yield 'unique objects' => [[new \stdClass(), new \stdClass()]],
+        ];
+    }
+
+    /**
+     * @dataProvider getInvalidValues
+     */
+    public function testInvalidValues($value)
+    {
+        $constraint = new Unique([
+            'message' => 'myMessage',
+        ]);
+        $this->validator->validate($value, $constraint);
+
+        $this->buildViolation('myMessage')
+             ->setParameter('{{ value }}', 'array')
+             ->setCode(Unique::IS_NOT_UNIQUE)
+             ->assertRaised();
+    }
+
+    public function getInvalidValues()
+    {
+        $object = new \stdClass();
+
+        return [
+            yield 'not unique booleans' => [[true, true]],
+            yield 'not unique integers' => [[1, 2, 3, 3]],
+            yield 'not unique floats' => [[0.1, 0.2, 0.1]],
+            yield 'not unique string' => [['a', 'b', 'a']],
+            yield 'not unique arrays' => [[[1, 1], [2, 3], [1, 1]]],
+            yield 'not unique objects' => [[$object, $object]],
+        ];
+    }
+}